rename files.
.

.

unwanted ring
